# Conflicts:
#	.idea/misc.xml

+++ app/src/main/java/com/caravan/caravan/ui/fragment/details/GuideDetailsFragment.kt
@@ -179,7 +179,7 @@ class GuideDetailsFragment : BaseFragment() {
             "Ogabek Matyakubov",
             true,
             4.5,
-            Price(150, "USD", "day"),
+            Price(150.0.toLong(), "USD", "day"),
             ArrayList<Language>().apply {
                 add(Language("1", "English", "Advanced"))
                 add(Language("2", "Uzbek", "Native"))
@@ -230,7 +230,7 @@ class GuideDetailsFragment : BaseFragment() {
                 add(Location("1", "Khorezm", "Khiva", "Ichan Qala"))
             },
             "Khiva in 3 days",
-            Price(1200, "USD", "trip"),
+            Price(1200.0.toLong(), "USD", "trip"),
             5, 10,
             guide,
             "+998997492581",
