+++ app/src/main/java/com/caravan/caravan/ui/fragment/main/SearchFragment.kt
@@ -15,6 +15,7 @@ import com.caravan.caravan.databinding.BottomDialogTripBinding
 import com.caravan.caravan.databinding.FragmentSearchBinding
 import com.caravan.caravan.model.*
 import com.caravan.caravan.ui.fragment.BaseFragment
+import com.caravan.caravan.utils.Extensions.toast
 
 
 class SearchFragment : BaseFragment() {
@@ -183,7 +184,7 @@ class SearchFragment : BaseFragment() {
                     "Ogabek Matyakubov",
                     true,
                     4.5,
-                    Price(150, "USD", "day"),
+                    Price(150.0.toLong(), "USD", "day"),
                     ArrayList<Language>().apply {
                         add(Language("1", "English", "Advanced"))
                         add(Language("1", "Uzbek", "Native"))
@@ -228,7 +229,7 @@ class SearchFragment : BaseFragment() {
             "Ogabek Matyakubov",
             true,
             4.5,
-            Price(150, "USD", "day"),
+            Price(150.0.toLong(), "USD", "day"),
             ArrayList<Language>().apply {
                 add(Language("1", "English", "Advanced"))
                 add(Language("1", "Uzbek", "Native"))
@@ -282,7 +283,7 @@ class SearchFragment : BaseFragment() {
                         add(Location("1", "Khorezm", "Khiva", "Ichan Qala"))
                     },
                     "Khiva in 3 days",
-                    Price(1200, "USD", "trip"),
+                    Price(1200.0.toLong(), "USD", "trip"),
                     5, 10,
                     guide,
                     "+998997492581",
